pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "h-cli"
version = "2.0.0"
description = "Terminal-based streaming client with themed UI and MPV playback"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "pyfiglet>=0.8",
    "Pillow>=9.0",
    "chafa.py>=1.2.0",
]

[project.scripts]
hcli = "hcli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hcli"]
